drfluffy/requirements.txt
Django==1.11.16
django_redis==4.10.0
requests==2.20.0
redis==3.0.1
djangorestframework==3.9.1
